Move repository directory into infrastructure directory and fix the directory name
