+++ main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"drello-api/pkg/infrastracture/mysql"
+	"drello-api/pkg/infrastructure/mysql"
 	"drello-api/pkg/presentation/rest"
 	"log"
 )
