+++ pkg/app/workspaces/getall.go
@@ -2,8 +2,8 @@ package workspaces
 
 import (
 	"context"
-	"drello-api/pkg/app/repository"
 	wdomain "drello-api/pkg/domain/workspace"
+	"drello-api/pkg/infrastructure/repository"
 )
 
 func GetAll(ctx context.Context, workspaceRepo repository.Workspace) (*GetAllOutput, error) {
